feat: add UI service module and testing HTML for modular components
- Implemented ui-service.js for handling UI interactions, including loading indicators, error messages, message history, and user input management.
- Added functions for showing/hiding option buttons and managing confirmation modals.
- Created test.html to validate module loading, DOM elements, and event bindings.
- Included basic styles and structure for testing interface.
