+++ static/js/dom-elements.js
@@ -1,0 +1,47 @@
+// DOM元素引用模块
+
+// 页面元素
+export const homePage = document.getElementById('homePage');
+export const chatPage = document.getElementById('chatPage');
+export const startButton = document.getElementById('startButton');
+export const exitButton = document.getElementById('exitButton');
+export const backButton = document.getElementById('backButton');
+
+// 对话元素
+export const characterName = document.getElementById('characterName');
+export const currentMessage = document.getElementById('currentMessage');
+export const messageInput = document.getElementById('messageInput');
+export const sendButton = document.getElementById('sendButton');
+export const playaudioButton = document.getElementById('playaudioButton');
+export const backgroundButton = document.getElementById('backgroundButton');
+export const historyButton = document.getElementById('historyButton');
+export const characterButton = document.getElementById('characterButton');
+export const continueButton = document.getElementById('continueButton');
+export const skipButton = document.getElementById('skipButton');
+export const micButton = document.getElementById('micButton');
+export const clickToContinue = document.getElementById('clickToContinue');
+export const optionButtonsContainer = document.getElementById('optionButtonsContainer');
+export const optionButtons = document.getElementById('optionButtons');
+
+// 历史记录模态框
+export const historyModal = document.getElementById('historyModal');
+export const historyMessages = document.getElementById('historyMessages');
+export const closeHistoryButton = document.getElementById('closeHistoryButton');
+
+// 角色选择模态框
+export const characterModal = document.getElementById('characterModal');
+export const characterList = document.getElementById('characterList');
+export const closeCharacterButton = document.getElementById('closeCharacterButton');
+
+// 确认对话框
+export const confirmModal = document.getElementById('confirmModal');
+export const confirmMessage = document.getElementById('confirmMessage');
+export const confirmYesButton = document.getElementById('confirmYesButton');
+export const confirmNoButton = document.getElementById('confirmNoButton');
+export const closeConfirmButton = document.getElementById('closeConfirmButton');
+
+// 加载和错误元素
+export const loadingIndicator = document.getElementById('loadingIndicator');
+export const errorContainer = document.getElementById('errorContainer');
+export const errorMessage = document.getElementById('errorMessage');
+export const errorCloseButton = document.getElementById('errorCloseButton');
