+++ static/js/main.js
@@ -1,1597 +1,129 @@
-const homePage = document.getElementById('homePage');
-
-const chatPage = document.getElementById('chatPage');
-
-const startButton = document.getElementById('startButton');
-
-const exitButton = document.getElementById('exitButton');
-
-const backButton = document.getElementById('backButton');
-
-
-
-// 对话元素
-
-const characterName = document.getElementById('characterName');
-
-const currentMessage = document.getElementById('currentMessage');
-
-const messageInput = document.getElementById('messageInput');
-
-const sendButton = document.getElementById('sendButton');
-
-// const clearButton = document.getElementById('clearButton'); // 已禁用：清空对话按钮
-const playaudioButton = document.getElementById('playaudioButton');
-
-const backgroundButton = document.getElementById('backgroundButton');
-
-const historyButton = document.getElementById('historyButton');
-
-const characterButton = document.getElementById('characterButton');
-
-const continueButton = document.getElementById('continueButton');
-
-const skipButton = document.getElementById('skipButton');
-
-const micButton = document.getElementById('micButton');
-
-const clickToContinue = document.getElementById('clickToContinue');
-const optionButtonsContainer = document.getElementById('optionButtonsContainer');
-const optionButtons = document.getElementById('optionButtons');
-
-const historyModal = document.getElementById('historyModal');
-
-const historyMessages = document.getElementById('historyMessages');
-
-const closeHistoryButton = document.getElementById('closeHistoryButton');
-
-const characterModal = document.getElementById('characterModal');
-
-const characterList = document.getElementById('characterList');
-
-const closeCharacterButton = document.getElementById('closeCharacterButton');
-
-
-
-// 确认对话框
-
-const confirmModal = document.getElementById('confirmModal');
-
-const confirmMessage = document.getElementById('confirmMessage');
-
-const confirmYesButton = document.getElementById('confirmYesButton');
-
-const confirmNoButton = document.getElementById('confirmNoButton');
-
-const closeConfirmButton = document.getElementById('closeConfirmButton');
-
-
-
-// 加载和错误元素
-
-const loadingIndicator = document.getElementById('loadingIndicator');
-
-const errorContainer = document.getElementById('errorContainer');
-
-const errorMessage = document.getElementById('errorMessage');
-
-const errorCloseButton = document.getElementById('errorCloseButton');
-
-
-
-// 角色数据
-
-let availableCharacters = [];
-
-let currentCharacter = null;
-
-let currentCharacterImages = []; // 当前角色的所有图片
-
-
-
-// 状态变量
-
-let isProcessing = false;
-
-let isPaused = false;
-
-let messageHistory = [];
-let currentTypingTimeout = null;
-
-let currentConfirmCallback = null;
-
-
-// 音频缓存对象和首次播放标记
-const audioCache = {};
-let firstAudioPlay = true;
-
-// 初始化
-
-document.addEventListener('DOMContentLoaded', () => {
-
-    // 加载角色数据
-
-    loadCharacters();
-
-    // 绑定页面切换事件
-
-    startButton.addEventListener('click', showChatPage);
-
-    backButton.addEventListener('click', confirmBackToHome);
-
-    exitButton.addEventListener('click', confirmExit);
-
-    // 绑定对话事件
-
-    sendButton.addEventListener('click', sendMessage);
-
-    // clearButton.addEventListener('click', clearChat); // 已禁用：清空对话按钮事件
-    playaudioButton.addEventListener('click', playAudio);
-
-    backgroundButton.addEventListener('click', changeBackground);
-
-    historyButton.addEventListener('click', toggleHistory);
-
-    closeHistoryButton.addEventListener('click', toggleHistory);
-
-    characterButton.addEventListener('click', toggleCharacterModal);
-
-    closeCharacterButton.addEventListener('click', toggleCharacterModal);
-
-    continueButton.addEventListener('click', continueOutput);
-
-    skipButton.addEventListener('click', skipTyping);
-
-    micButton.addEventListener('click', toggleRecording);
-
-    errorCloseButton.addEventListener('click', hideError);
-
-    // 绑定确认对话框事件
-
-    confirmYesButton.addEventListener('click', handleConfirmYes);
-
-    confirmNoButton.addEventListener('click', handleConfirmNo);
-
-    closeConfirmButton.addEventListener('click', hideConfirmModal);
-
-    // 绑定回车键发送消息
-
-    messageInput.addEventListener('keydown', (e) => {
-
-        if (e.key === 'Enter' && !e.shiftKey) {
-
-            e.preventDefault();
-
-            sendMessage();
-
-        }
-
-    });
-
-    // 绑定点击事件继续输出
-
-    currentMessage.addEventListener('click', continueOutput);
-
-    clickToContinue.addEventListener('click', continueOutput);
-
+// 主入口文件 - 事件绑定和初始化
+import { 
+    startButton, 
+    backButton, 
+    exitButton, 
+    sendButton, 
+    playaudioButton, 
+    backgroundButton, 
+    historyButton, 
+    closeHistoryButton, 
+    characterButton, 
+    closeCharacterButton, 
+    continueButton, 
+    skipButton, 
+    micButton, 
+    errorCloseButton, 
+    confirmYesButton, 
+    confirmNoButton, 
+    closeConfirmButton, 
+    messageInput,
+    currentMessage,
+    clickToContinue
+} from './dom-elements.js';
+
+import { 
+    showChatPage, 
+    confirmBackToHome, 
+    confirmExit, 
+    toggleHistory, 
+    hideError, 
+    handleConfirmYes, 
+    handleConfirmNo, 
+    hideConfirmModal,
+    showOptionButtons,
+    showError
+} from './ui-service.js';
+
+import { 
+    loadCharacters, 
+    toggleCharacterModal, 
+    getCurrentCharacter 
+} from './character-service.js';
+
+import { 
+    playAudio, 
+    toggleRecording,
+    stopCurrentAudio
+} from './audio-service.js';
+
+import { 
+    sendMessage, 
+    changeBackground, 
+    continueOutput, 
+    skipTyping 
+} from './chat-service.js';
+
+// 全局错误处理
+window.addEventListener('error', (event) => {
+    console.error('全局错误:', event.error);
+    showError(`发生错误: ${event.error.message}`);
 });
 
-// 流式处理器实例
-let streamProcessor = null;
+window.addEventListener('unhandledrejection', (event) => {
+    console.error('未处理的Promise拒绝:', event.reason);
+    showError(`请求失败: ${event.reason}`);
+});
 
-// 发送消息
-
-async function sendMessage() {
-
-    // 获取消息内容
-
-    const message = messageInput.value.trim();
-
-    // 检查消息是否为空
-
-    if (!message) {
-
-        showError('请输入消息');
-
-        return;
-
-    }
-
-    // 检查是否正在处理请求
-
-    if (isProcessing) {
-
-        showError('正在处理上一条消息，请稍候');
-
-        return;
-
-    }
-
-    // 创建新的流式处理器
-    streamProcessor = new StreamProcessor();
-
-    // 跟踪已添加到历史记录的内容长度
-    let addedToHistoryLength = 0;
-
-    // 设置回调函数
-    streamProcessor.setCallbacks(
-        // 字符回调 - 每个字符输出时调用
-        (fullContent) => {
-            // 只显示新增的内容，不显示已添加到历史记录的部分
-            const newContent = fullContent.substring(addedToHistoryLength);
-            updateCurrentMessage('assistant', newContent, true);
-        },
-        // 暂停回调
-        (fullContent) => {
-            isPaused = true;
-            const newContent = fullContent.substring(addedToHistoryLength);
-            if (newContent) {
-                const characterName = currentCharacter ? currentCharacter.name : 'AI助手';
-                prefetchAudio(newContent, characterName, () => {
-                    addToHistory('assistant', newContent, characterName);
-                    updateCurrentMessage('assistant', newContent);
-                    addedToHistoryLength = fullContent.length;
-                    showContinuePrompt();
-                    playAudio();
-                });
-            } else {
-                showContinuePrompt();
-                playAudio();
-            }
-        },
-        // 完成回调
-        (fullContent) => {
-            const remainingContent = fullContent.substring(addedToHistoryLength);
-            if (remainingContent) {
-                const characterName = currentCharacter ? currentCharacter.name : 'AI助手';
-                prefetchAudio(remainingContent, characterName, () => {
-                    addToHistory('assistant', remainingContent, characterName);
-                    updateCurrentMessage('assistant', remainingContent);
-                    hideContinuePrompt();
-                    enableUserInput();
-                    isPaused = false;
-                    if (window.pendingOptions && window.pendingOptions.length > 0) {
-                        showOptionButtons(window.pendingOptions);
-                        window.pendingOptions = null;
-                    }
-                    playAudio();
-                });
-            } else {
-                hideContinuePrompt();
-                enableUserInput();
-                isPaused = false;
-                if (window.pendingOptions && window.pendingOptions.length > 0) {
-                    showOptionButtons(window.pendingOptions);
-                    window.pendingOptions = null;
-                }
-                playAudio();
-            }
-        }
-    );
-
-    // 更新当前消息为用户消息
-    updateCurrentMessage('user', message);
-
-    // 添加到历史记录
-    addToHistory('user', message);
-
-    // 隐藏选项按钮
-    hideOptionButtons();
-
-    // 清空输入框
-    messageInput.value = '';
-
-    // 禁用用户输入
-    disableUserInput();
-
+// 初始化
+document.addEventListener('DOMContentLoaded', () => {
     try {
-        // 发送API请求
-        const response = await fetch('/api/chat/stream', {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json'
-            },
-            body: JSON.stringify({ message })
-        });
-
-        // 检查响应状态
-        if (!response.ok) {
-            const errorData = await response.json();
-            throw new Error(errorData.error || '请求失败');
-        }
-
-        // 获取响应的读取器
-        const reader = response.body.getReader();
-        const decoder = new TextDecoder();
-
-        // 准备接收流式响应
-        updateCurrentMessage('assistant', '\n');
-
-        // 读取流式响应
-        while (true) {
-            const { done, value } = await reader.read();
-
-            if (done) {
-                break;
-            }
-
-            // 解码响应数据
-            const chunk = decoder.decode(value, { stream: true });
-
-            try {
-                // 解析流式数据
-                const lines = chunk.split('\n').filter(line => line.trim());
-
-                for (const line of lines) {
-                    if (line.startsWith('data: ')) {
-                        const jsonStr = line.slice(6);
-
-                        if (jsonStr === '[DONE]') {
-                            // 标记流式处理结束
-                            streamProcessor.markEnd();
-                            break;
-                        }
-
-                        try {
-                            const data = JSON.parse(jsonStr);
-
-                            // 处理错误
-                            if (data.error) {
-                                throw new Error(data.error);
-                            }
-
-                            // 处理mood字段 - 立即处理表情变化
-                            if (data.mood !== undefined) {
-                                console.log('收到mood数据:', data.mood);
-                                // 立即并行处理表情变化
-                                handleMoodChange(data.mood);
-                            }
-
-                            // 处理content字段
-                            if (data.content) {
-                                // 将数据添加到流式处理器
-                                streamProcessor.addData(data.content);
-                            }
-
-                            // 处理选项数据
-                            if (data.options) {
-                                // 存储选项数据，等待输出完成后显示
-                                window.pendingOptions = data.options;
-                            }
-                        } catch (e) {
-                            console.error('解析JSON失败:', e, jsonStr);
-                        }
-                    }
-                }
-            } catch (e) {
-                console.error('解析流式响应失败:', e);
-            }
-        }
-
-    } catch (error) {
-        console.error('发送消息失败:', error);
-        showError(`发送消息失败: ${error.message}`);
-        hideLoading();
-        enableUserInput();
-
-        // 重置流式处理器
-        if (streamProcessor) {
-            streamProcessor.reset();
-        }
-    }
-
-}
-
-
-
-// 清空对话 - 已禁用但保留功能代码
-/*
-async function clearChat() {
-
-    // 检查是否正在处理请求
-
-    if (isProcessing) {
-
-        showError('正在处理请求，请稍候');
-
-        return;
-
-    }
-
-    // 显示加载指示器
-
-    showLoading();
-
-    try {
-
-        // 发送API请求
-
-        const response = await fetch('/api/clear', {
-
-            method: 'POST',
-
-            headers: {
-
-                'Content-Type': 'application/json'
-
-            },
-
-            body: JSON.stringify({ prompt_type: 'friendly' })
-
-        });
-
-        // 解析响应
-
-        const data = await response.json();
-
-        // 检查响应是否成功
-
-        if (!data.success) {
-
-            throw new Error(data.error || '请求失败');
-
-        }
-
-        // 清空历史记录
-
-        messageHistory = [];
-
-        historyMessages.innerHTML = '';
-
-        // 更新当前消息
-
-        updateCurrentMessage('assistant', '对话已重置。有什么我可以帮助您的？');
-
-    } catch (error) {
-
-        console.error('清空对话失败:', error);
-
-        showError(`清空对话失败: ${error.message}`);
-
-    } finally {
-
-        // 隐藏加载指示器
-
-        hideLoading();
-
-    }
-
-}
-*/
-var first = 0; // 用于调试
-async function playAudio() {
-    if (isProcessing) {
-        showError('正在处理请求，请稍候');
-        return;
-    }
-
-    const text = currentMessage.textContent.trim();
-    if (!text) {
-        showError('无法朗读空内容');
-        return;
-    }
-
-    // 判断缓存
-    let audioBlob = audioCache[text];
-    if (audioBlob) {
-        console.log('[playAudio] 命中缓存:', text);
-        try {
-            const audio = new Audio();
-            const url = URL.createObjectURL(audioBlob);
-            audio.src = url;
-            audio.onended = () => { URL.revokeObjectURL(url); };
-            audio.onerror = () => { showError('音频播放失败'); URL.revokeObjectURL(url); };
-            window.currentAudio = audio; // 保存当前音频对象
-            await audio.play();
-        } catch (error) {
-            console.error('播放音频失败:', error);
-            showError(`播放失败: ${error.message}`);
-        }
-        return;
-    } else {
-        console.log('[playAudio] 未命中缓存，重新合成:', text);
-        showLoading(); // 未命中缓存时显示处理弹窗
-    }
-
-    // 首次播放需要等待和弹窗
-    if (firstAudioPlay) {
-        firstAudioPlay = false;
-        alert('首次合成可能需要较长时间，请耐心等待。');
-        // showLoading(); // 已在未命中缓存时显示，无需重复
-    }
-
-    try {
-        const response = await fetch('/api/tts', {
-            method: 'POST',
-            headers: { 'Content-Type': 'application/json' },
-            body: JSON.stringify({
-                text: text,
-                role: currentCharacter ? currentCharacter.name : 'AI助手'
-            })
-        });
-
-        if (!response.ok) {
-            const errData = await response.json().catch(() => ({}));
-            throw new Error(errData.error || '获取音频失败');
-        }
-
-        const blob = await response.blob();
-        if (blob.size === 0) {
-            throw new Error('收到的音频数据为空');
-        }
-        // 缓存音频
-        audioCache[text] = blob;
-
-        const audio = new Audio();
-        const url = URL.createObjectURL(blob);
-        audio.src = url;
-        audio.onended = () => { URL.revokeObjectURL(url); };
-        audio.onerror = () => { showError('音频播放失败'); URL.revokeObjectURL(url); };
-        window.currentAudio = audio; // 保存当前音频对象
-        await audio.play();
-    } catch (error) {
-        console.error('播放音频失败:', error);
-        showError(`播放失败: ${error.message}`);
-    } finally {
-        hideLoading(); // 合成完成后隐藏处理弹窗
-    }
-}
-
-
-
-// 更换背景
-
-async function changeBackground() {
-
-    // 检查是否正在处理请求
-
-    if (isProcessing) {
-
-        showError('正在处理请求，请稍候');
-
-        return;
-
-    }
-
-    // 显示加载指示器
-
-    showLoading();
-
-    try {
-
-        // 发送API请求
-
-        const response = await fetch('/api/background', {
-
-            method: 'POST',
-
-            headers: {
-
-                'Content-Type': 'application/json'
-
-            },
-
-            body: JSON.stringify({})
-
-        });
-
-        // 解析响应
-
-        const data = await response.json();
-
-        // 检查响应是否成功
-
-        if (!data.success) {
-
-            throw new Error(data.error || '请求失败');
-
-        }
-
-        // 更新背景图片
-
-        if (data.background_url) {
-
-            updateBackground(data.background_url);
-
-            // 添加系统消息
-            const promptMessage = data.prompt ?
-                `背景已更新，提示词: "${data.prompt}"` :
-                '背景已更新';
-
-            updateCurrentMessage('system', promptMessage);
-
-        }
-
-    } catch (error) {
-
-        console.error('更换背景失败:', error);
-
-        showError(`更换背景失败: ${error.message}`);
-
-    } finally {
-
-        // 隐藏加载指示器
-
-        hideLoading();
-
-    }
-
-}
-
-
-
-// 更新当前消息
-function updateCurrentMessage(role, content, isStreaming = false) {
-    // 如果不是流式输出，停止当前正在进行的打字效果
-
-    if (!isStreaming && currentTypingTimeout) {
-
-        clearTimeout(currentTypingTimeout);
-
-        currentTypingTimeout = null;
-
-    }
-
-    // 更新角色名称
-
-    if (role === 'user') {
-
-        characterName.textContent = '你';
-
-        characterName.style.color = '#90caf9';
-
-    } else if (role === 'assistant') {
-
-        // 使用当前角色名称
-
-        if (currentCharacter) {
-
-            characterName.textContent = currentCharacter.name;
-
-            characterName.style.color = currentCharacter.color;
-
-        } else {
-            characterName.textContent = 'AI';
-            characterName.style.color = '#ffeb3b';
-
-        }
-
-    } else if (role === 'system') {
-
-        characterName.textContent = '系统';
-
-        characterName.style.color = '#4caf50';
-
-    }
-
-    // 直接更新内容（JSON格式下不需要过滤【】内容）
-    currentMessage.textContent = content;
-
-}
-
-
-
-
-
-
-
-
-
-// 添加消息到历史记录
-function addToHistory(role, content, customName = null) {
-    // JSON格式下不需要过滤【】标记
-
-    // 添加到内存中的历史记录
-
-    messageHistory.push({
-
-        role: role === 'assistant_continue' ? 'assistant' : role,
-
-        content: content
-
-    });
-
-    // 添加到历史记录面板
-
-    const messageDiv = document.createElement('div');
-
-    messageDiv.className = `history-message history-${role}`;
-
-    const roleSpan = document.createElement('div');
-
-    roleSpan.className = 'history-role';
-
-    // 使用当前角色名称或自定义名称
-    let roleName = '';
-    if (role === 'user') {
-        roleName = '你';
-        roleSpan.textContent = roleName;
-    } else if (role === 'assistant') {
-        roleName = customName || (currentCharacter ? currentCharacter.name : 'AI');
-        roleSpan.textContent = roleName;
-    } else {
-        roleName = '系统';
-        roleSpan.textContent = roleName;
-    }
-
-    const contentDiv = document.createElement('div');
-
-    contentDiv.className = 'history-content';
-
-    // 为每个段落添加角色名称前缀
-    if (role === 'assistant') {
-        contentDiv.textContent = content;
-    } else {
-        contentDiv.textContent = content;
-    }
-
-    messageDiv.appendChild(roleSpan);
-
-    messageDiv.appendChild(contentDiv);
-
-    historyMessages.appendChild(messageDiv);
-
-}
-
-
-
-// 切换历史记录面板
-
-function toggleHistory() {
-
-    if (historyModal.style.display === 'flex') {
-
-        historyModal.style.display = 'none';
-
-    } else {
-
-        historyModal.style.display = 'flex';
-
-        // 滚动到底部
-
-        historyMessages.scrollTop = historyMessages.scrollHeight;
-
-    }
-
-}
-
-
-
-// 更新背景图片
-
-function updateBackground(url) {
-
-    // 获取背景元素
-
-    const backgroundElements = document.getElementsByClassName('background-image');
-
-    if (backgroundElements.length > 0) {
-
-        const backgroundElement = backgroundElements[0];
-
-        // 创建新背景元素
-
-        const newBackground = document.createElement('div');
-
-        newBackground.className = 'background-image';
-
-        newBackground.style.backgroundImage = `url('${url}')`;
-
-        newBackground.style.opacity = '0';
-
-        // 添加新背景
-
-        backgroundElement.parentNode.appendChild(newBackground);
-
-        // 淡入新背景
-
-        setTimeout(() => {
-
-            newBackground.style.opacity = '1';
-
-            // 淡出旧背景
-
-            backgroundElement.style.opacity = '0';
-
-            // 移除旧背景
-
-            setTimeout(() => {
-
-                backgroundElement.remove();
-
-            }, 1000);
-
-        }, 10);
-
-    }
-
-}
-
-
-
-// 显示加载指示器
-
-function showLoading() {
-
-    isProcessing = true;
-
-    loadingIndicator.style.display = 'flex';
-
-}
-
-
-
-// 隐藏加载指示器
-
-function hideLoading() {
-
-    isProcessing = false;
-
-    loadingIndicator.style.display = 'none';
-
-}
-
-
-
-// 显示错误信息
-
-function showError(message) {
-
-    errorMessage.textContent = message;
-
-    errorContainer.style.display = 'block';
-
-}
-
-
-
-// 隐藏错误信息
-
-function hideError() {
-
-    errorContainer.style.display = 'none';
-
-}
-
-
-
-// 加载角色数据
-
-async function loadCharacters() {
-
-    try {
-
-        // 显示加载指示器
-
-        showLoading();
-
-        // 发送API请求
-
-        const response = await fetch('/api/characters');
-
-        // 解析响应
-
-        const data = await response.json();
-
-        // 检查响应是否成功
-
-        if (!data.success) {
-
-            throw new Error(data.error || '请求失败');
-
-        }
-
-        // 更新角色数据
-
-        availableCharacters = data.available_characters;
-
-        currentCharacter = data.current_character;
-
-        // 调试：输出角色信息
-        console.log('当前角色信息:', currentCharacter);
-
-        // 更新角色图片
-
-        updateCharacterImage();
-
-        // 更新当前消息
-
-        if (currentCharacter && currentCharacter.welcome) {
-
-            updateCurrentMessage('assistant', currentCharacter.welcome);
-
-        } else {
-
-            updateCurrentMessage('assistant', '欢迎使用CABM！我是您的AI助手，请输入消息开始对话。');
-
-        }
-
-    } catch (error) {
-
-        console.error('加载角色数据失败:', error);
-
-        showError(`加载角色数据失败: ${error.message}`);
-
-        // 设置默认消息
-
-        updateCurrentMessage('assistant', '欢迎使用CABM！我是您的AI助手，请输入消息开始对话。');
-
-    } finally {
-
-        // 隐藏加载指示器
-
-        hideLoading();
-
-    }
-
-}
-
-
-
-// 更新角色图片
-
-function updateCharacterImage() {
-    // 获取角色图片元素和容器
-    const characterImage = document.getElementById('characterImage');
-    const characterContainer = document.querySelector('.character-container');
-
-    // 如果有当前角色，更新图片
-    if (currentCharacter && characterImage) {
-        // 加载角色的所有图片
-        loadCharacterImages(currentCharacter.id);
-
-        // 根据CALIB值调整位置
-        if (characterContainer && typeof currentCharacter.calib !== 'undefined') {
-            // 基准值为50%，根据CALIB值进行调整
-            const baseTop = 50;
-            const adjustedTop = baseTop + currentCharacter.calib;
-            console.log(`角色位置调整: CALIB=${currentCharacter.calib}, 调整后位置=${adjustedTop}%`);
-            characterContainer.style.top = `${adjustedTop}%`;
-        } else if (characterContainer) {
-            // 如果没有CALIB值，使用默认位置
-            console.log('使用默认位置: 50%');
-            characterContainer.style.top = '50%';
-        }
-    }
-}
-
-
-
-// 切换角色选择弹窗
-
-function toggleCharacterModal() {
-
-    if (characterModal.style.display === 'flex') {
-
-        characterModal.style.display = 'none';
-
-    } else {
-
-        // 显示角色选择弹窗
-
-        characterModal.style.display = 'flex';
-
-        // 渲染角色列表
-
-        renderCharacterList();
-
-    }
-
-}
-
-
-
-// 渲染角色列表
-
-function renderCharacterList() {
-
-    // 清空角色列表
-
-    characterList.innerHTML = '';
-
-    // 遍历可用角色
-
-    availableCharacters.forEach(character => {
-
-        // 创建角色卡片
-
-        const card = document.createElement('div');
-
-        card.className = `character-card ${currentCharacter && character.id === currentCharacter.id ? 'active' : ''}`;
-
-        card.dataset.characterId = character.id;
-
-        // 创建角色图片
-
-        const imageContainer = document.createElement('div');
-
-        imageContainer.className = 'character-card-image';
-
-        const image = document.createElement('img');
-
-        image.src = character.image.endsWith('/')
-            ? `${character.image}1.png`
-            : `${character.image}/1.png`;
-
-        image.alt = character.name;
-
-        imageContainer.appendChild(image);
-
-        // 创建角色名称
-
-        const name = document.createElement('div');
-
-        name.className = 'character-card-name';
-
-        name.textContent = character.name;
-
-        name.style.color = character.color;
-
-        // 创建角色描述
-
-        const description = document.createElement('div');
-
-        description.className = 'character-card-description';
-
-        description.textContent = character.description;
-
-        // 组装角色卡片
-
-        card.appendChild(imageContainer);
-
-        card.appendChild(name);
-
-        card.appendChild(description);
-
-        // 添加点击事件
-
-        card.addEventListener('click', () => {
-
-            selectCharacter(character.id);
-
-        });
-
-        // 添加到角色列表
-
-        characterList.appendChild(card);
-
-    });
-
-}
-
-
-
-// 选择角色
-
-async function selectCharacter(characterId) {
-
-    try {
-
-        // 显示加载指示器
-
-        showLoading();
-
-        // 发送API请求
-
-        const response = await fetch(`/api/characters/${characterId}`, {
-
-            method: 'POST',
-
-            headers: {
-
-                'Content-Type': 'application/json'
-
-            }
-
-        });
-
-        // 解析响应
-
-        const data = await response.json();
-
-        // 检查响应是否成功
-
-        if (!data.success) {
-
-            throw new Error(data.error || '请求失败');
-
-        }
-
-        // 更新当前角色
-
-        currentCharacter = data.character;
-
-        // 调试：输出角色信息
-        console.log('切换后的角色信息:', currentCharacter);
-
-        // 更新角色图片
-
-        updateCharacterImage();
-
-        // 更新角色列表
-
-        renderCharacterList();
-
-        // 更新当前消息
-
-        if (currentCharacter.welcome) {
-
-            updateCurrentMessage('assistant', currentCharacter.welcome);
-
-        }
-
-        // 隐藏选项按钮
-        hideOptionButtons();
-
-        // 关闭角色选择弹窗
-
-        toggleCharacterModal();
-
-    } catch (error) {
-
-        console.error('选择角色失败:', error);
-
-        showError(`选择角色失败: ${error.message}`);
-
-    } finally {
-
-        // 隐藏加载指示器
-
-        hideLoading();
-
-    }
-
-}
-
-
-
-// 页面切换函数
-
-function showHomePage() {
-
-    homePage.classList.add('active');
-
-    chatPage.classList.remove('active');
-
-}
-
-
-
-function showChatPage() {
-
-    homePage.classList.remove('active');
-
-    chatPage.classList.add('active');
-
-}
-
-
-
-// 确认返回主页
-
-function confirmBackToHome() {
-
-    // 如果有未完成的对话，显示确认对话框
-
-    if (messageHistory.length > 0) {
-
-        showConfirmModal('确定要返回吗？', () => {
-
-            // 清空对话历史
-
-            // messageHistory = [];
-
-            // historyMessages.innerHTML = '';
-
-            // 返回主页
-
-            showHomePage();
-
-        });
-
-    } else {
-
-        // 直接返回主页
-
-        showHomePage();
-
-    }
-
-}
-
-
-
-// 确认退出应用
-
-function confirmExit() {
-
-    showConfirmModal('确定要退出应用吗？', exitApplication);
-
-}
-
-
-
-// 退出应用
-
-async function exitApplication() {
-
-    try {
-
-        // 显示加载指示器
-
-        showLoading();
-
-        // 发送API请求
-
-        const response = await fetch('/api/exit', {
-
-            method: 'POST',
-
-            headers: {
-
-                'Content-Type': 'application/json'
-
-            }
-
-        });
-
-        // 解析响应
-
-        const data = await response.json();
-
-        // 检查响应是否成功
-
-        if (!data.success) {
-
-            throw new Error(data.error || '请求失败');
-
-        }
-
-        // 关闭窗口
-
-        window.close();
-
-        // 如果window.close()不起作用（大多数浏览器会阻止），显示提示
-
-        setTimeout(() => {
-
-            showError('请手动关闭浏览器窗口');
-
-        }, 1000);
-
-    } catch (error) {
-
-        console.error('退出应用失败:', error);
-
-        showError(`退出应用失败: ${error.message}`);
-
-        hideLoading();
-
-    }
-
-}
-
-
-
-// 显示确认对话框
-
-function showConfirmModal(message, callback) {
-
-    confirmMessage.textContent = message;
-
-    currentConfirmCallback = callback;
-
-    confirmModal.style.display = 'flex';
-
-}
-
-
-
-// 隐藏确认对话框
-
-function hideConfirmModal() {
-
-    confirmModal.style.display = 'none';
-
-    currentConfirmCallback = null;
-
-}
-
-
-
-// 处理确认对话框的确定按钮
-
-function handleConfirmYes() {
-
-    if (currentConfirmCallback) {
-
-        currentConfirmCallback();
-
-    }
-
-    hideConfirmModal();
-
-}
-
-
-
-// 处理确认对话框的取消按钮
-
-function handleConfirmNo() {
-
-    hideConfirmModal();
-
-}
-
-
-
-// 继续输出
-
-function continueOutput() {
-    // 自动中断正在播放的语音
-    if (window.currentAudio && typeof window.currentAudio.pause === 'function') {
-        window.currentAudio.pause();
-        window.currentAudio.currentTime = 0;
-        window.currentAudio = null;
-    }
-    console.log('continueOutput called, isPaused:', isPaused, 'streamProcessor:', streamProcessor);
-
-    if (isPaused && streamProcessor) {
-        isPaused = false;
-        hideContinuePrompt();
-
-        // 继续流式处理
-        streamProcessor.continue();
-    } else if (streamProcessor) {
-        // 如果没有暂停但有流式处理器，也尝试继续
-        hideContinuePrompt();
-        streamProcessor.continue();
-    }
-}
-
-// 跳过打字效果
-function skipTyping() {
-    if (streamProcessor && streamProcessor.isProcessing()) {
-        // 跳过当前流式处理
-        streamProcessor.skip();
-
-        // 重置状态
-        isPaused = false;
-        hideContinuePrompt();
-        enableUserInput();
-    }
-}
-
-let recognition; // 全局变量存储语音识别实例
-var first_mic = 0;
-
-function toggleRecording() {
-    if (first_mic === 0) {
-        first_mic = 1;
-        alert('请确保你访问的地址为本地地址或https协议地址，否则浏览器可能会阻止调用麦克风！！！');
-    }
-    let isRecording = micButton.classList.toggle('recording');
-    if (isRecording) {
-        if (!recognition) {
-            const SpeechRecognition = window.SpeechRecognition || 
-                                    window.webkitSpeechRecognition ||
-                                    window.mozSpeechRecognition || 
-                                    window.msSpeechRecognition;
-            
-            if (!SpeechRecognition) {
-                console.error('当前浏览器不支持语音识别API');
-                showError('您的浏览器不支持语音识别功能');
-                micButton.classList.remove('recording');
-                return;
-            }
-
-            recognition = new SpeechRecognition();
-            recognition.lang = 'zh-CN';
-            recognition.interimResults = false;
-            
-            recognition.onresult = (event) => {
-                const transcript = event.results[0][0].transcript.trim();
-                if (transcript) {
-                    messageInput.value += transcript;
-                }
-            };
-            
-            recognition.onerror = (event) => {
-                console.error('语音识别错误:', event.error);
-                showError(`语音识别错误: ${event.error}`);
-                micButton.classList.remove('recording');
-            };
-            
-            recognition.onend = () => {
-                if (micButton.classList.contains('recording')) {
-                    micButton.classList.remove('recording');
-                }
-            };
-        }
+        console.log('开始初始化CABM应用...');
         
-        try {
-            recognition.start();
-        } catch (e) {
-            console.error('启动语音识别失败:', e);
-            showError('无法启动语音识别');
-            micButton.classList.remove('recording');
-        }
-    } else {
-        if (recognition) {
-            recognition.stop();
-        }
-    }
-}
+        // 加载角色数据
+        loadCharacters();
 
-// 禁用用户输入
-function disableUserInput() {
-    messageInput.disabled = true;
-    sendButton.disabled = true;
-    micButton.disabled = true;
-    messageInput.placeholder = "AI正在回复中...";
-}
+        // 绑定页面切换事件
+        startButton.addEventListener('click', showChatPage);
+        backButton.addEventListener('click', confirmBackToHome);
+        exitButton.addEventListener('click', confirmExit);
 
-// 启用用户输入
-function enableUserInput() {
-    messageInput.disabled = false;
-    sendButton.disabled = false;
-    micButton.disabled = false;
-    messageInput.placeholder = "输入消息...";
-}
+        // 绑定对话事件
+        sendButton.addEventListener('click', sendMessage);
+        playaudioButton.addEventListener('click', () => playAudio(getCurrentCharacter(), false)); // 用户主动播放
+        backgroundButton.addEventListener('click', changeBackground);
+        historyButton.addEventListener('click', toggleHistory);
+        closeHistoryButton.addEventListener('click', toggleHistory);
+        characterButton.addEventListener('click', toggleCharacterModal);
+        closeCharacterButton.addEventListener('click', toggleCharacterModal);
+        continueButton.addEventListener('click', continueOutput);
+        skipButton.addEventListener('click', skipTyping);
+        micButton.addEventListener('click', () => toggleRecording(messageInput, micButton, showError));
+        errorCloseButton.addEventListener('click', hideError);
 
-// 显示"点击屏幕继续"提示
-function showContinuePrompt(promptText = '▽') {
-    // 激活继续按钮来显示提示
-    continueButton.classList.add('active');
+        // 绑定确认对话框事件
+        confirmYesButton.addEventListener('click', handleConfirmYes);
+        confirmNoButton.addEventListener('click', handleConfirmNo);
+        closeConfirmButton.addEventListener('click', hideConfirmModal);
 
-    // 如果有clickToContinue元素，显示它
-    if (clickToContinue) {
-        clickToContinue.style.display = 'block';
-        clickToContinue.textContent = promptText;
-    }
-
-    // 移除之前的监听器（如果存在）
-    if (currentScreenClickHandler) {
-        document.removeEventListener('click', currentScreenClickHandler);
-    }
-
-    // 创建新的点击监听器
-    currentScreenClickHandler = (e) => {
-        // 检查点击的目标是否是按钮或其他交互元素
-        const clickedElement = e.target;
-
-        // 如果点击的是按钮、输入框或其他交互元素，不触发继续
-        if (clickedElement.tagName === 'BUTTON' ||
-            clickedElement.tagName === 'INPUT' ||
-            clickedElement.tagName === 'TEXTAREA' ||
-            clickedElement.tagName === 'IMG' ||
-            clickedElement.classList.contains('modal') ||
-            clickedElement.classList.contains('btn') ||
-            clickedElement.classList.contains('continue-prompt') ||
-            clickedElement.closest('button') ||
-            clickedElement.closest('.btn') ||
-            clickedElement.closest('.modal') ||
-            clickedElement.closest('.history-modal') ||
-            clickedElement.closest('.character-modal') ||
-            clickedElement.closest('.confirm-modal') ||
-            clickedElement.closest('.control-buttons') ||
-            clickedElement.closest('.user-input-container') ||
-            clickedElement.closest('.chat-header') ||
-            clickedElement.closest('.character-container') ||
-            clickedElement.closest('.continue-prompt')) {
-            return; // 不处理按钮等元素的点击
-        }
-
-        // 只有点击对话区域或背景区域才触发继续
-        if (clickedElement.closest('.dialog-container') ||
-            clickedElement.closest('.background-container') ||
-            clickedElement === document.body ||
-            clickedElement.classList.contains('page')) {
-            console.log('Screen clicked during pause');
-            continueOutput();
-        }
-    };
-
-    // 延迟添加监听器，避免立即触发
-    setTimeout(() => {
-        document.addEventListener('click', currentScreenClickHandler);
-    }, 100);
-}
-
-// 全局变量来存储当前的点击监听器
-let currentScreenClickHandler = null;
-
-// 显示选项按钮
-function showOptionButtons(options) {
-    // 清空现有选项
-    optionButtons.innerHTML = '';
-
-    // 创建选项按钮
-    options.forEach((option, index) => {
-        const button = document.createElement('button');
-        button.className = 'option-button';
-        button.textContent = option;
-        button.addEventListener('click', () => {
-            selectOption(option);
+        // 绑定回车键发送消息
+        messageInput.addEventListener('keydown', (e) => {
+            if (e.key === 'Enter' && !e.shiftKey) {
+                e.preventDefault();
+                sendMessage();
+            }
         });
-        optionButtons.appendChild(button);
-    });
 
-    // 显示选项容器
-    optionButtonsContainer.classList.add('show');
-}
-
-// 隐藏选项按钮
-function hideOptionButtons() {
-    optionButtonsContainer.classList.remove('show');
-    optionButtons.innerHTML = '';
-}
-
-// 选择选项
-function selectOption(option) {
-    // 隐藏选项按钮
-    hideOptionButtons();
-
-    // 将选项内容设置到输入框
-    messageInput.value = option;
-
-    // 自动发送消息
-    sendMessage();
-}
-
-// **处理提取的【】内容的函数**
-// ⚠这个函数似乎没有被使用，但最好别删
-// function handleExtractedBracketContents(contents) {
-//     // 这里是处理提取内容的地方，你可以在这里添加自定义逻辑
-//     console.log('处理提取的【】内容:', contents);
-
-//     // 示例：解析心情状态
-//     contents.forEach((content, index) => {
-//         console.log(`第${index + 1}个【】内容:`, content);
-
-//         // 如果内容是数字，可能是心情状态
-//         if (/^\d+$/.test(content.trim())) {
-//             const moodNumber = parseInt(content.trim());
-//             const moods = ['', '平静', '兴奋', '愤怒', '失落'];
-//             if (moodNumber >= 1 && moodNumber <= 4) {
-//                 console.log(`检测到心情状态: ${moodNumber} - ${moods[moodNumber]}`);
-//                 // 在这里可以根据心情状态做相应处理，比如改变UI样式等
-//             }
-//         }
-//     });
-
-//     // 你可以在这里添加更多的处理逻辑
-//     // 例如：存储到全局变量、发送到服务器、更新UI等
-// }
-
-// 隐藏"点击屏幕继续"提示
-function hideContinuePrompt() {
-    // 隐藏点击继续提示
-    if (clickToContinue) {
-        clickToContinue.style.display = 'none';
+        // 绑定点击事件继续输出
+        currentMessage.addEventListener('click', continueOutput);
+        clickToContinue.addEventListener('click', continueOutput);
+        
+        console.log('CABM应用初始化完成');
+    } catch (error) {
+        console.error('初始化失败:', error);
+        showError(`初始化失败: ${error.message}`);
     }
+});
 
-    // 取消激活继续按钮
-    continueButton.classList.remove('active');
-
-    // 移除全屏点击监听器
-    if (currentScreenClickHandler) {
-        document.removeEventListener('click', currentScreenClickHandler);
-        currentScreenClickHandler = null;
-    }
-}
-
-
+// 注册快捷键
 function registrationShortcuts(config) {
     document.addEventListener('keydown', e => {
         if (e.key in config) {
-            // For single-letter keys, require Alt to be pressed
             if (e.key.length === 1) {
                 if (e.altKey) {
                     config[e.key]();
                 }
-            }
-            else {
+            } else {
                 config[e.key]();
             }
         }
@@ -1602,145 +134,12 @@ registrationShortcuts({
     Enter: continueOutput,
     s: skipTyping,
     h: toggleHistory,
-    b: changeBackground,
-    // c: clearChat, // 已禁用：清空对话快捷键
-    // ç: clearChat // 已禁用：清空对话快捷键（为了解决部分快捷键冲突）
+    b: changeBackground
 });
 
-// 处理mood变化（新的JSON格式）
-function handleMoodChange(moodNumber) {
-    console.log('处理mood变化:', moodNumber);
-
-    // 确保mood是有效数字
-    const imageNumber = parseInt(moodNumber);
-
-    if (!isNaN(imageNumber) && imageNumber > 0) {
-        // 如果是有效数字，切换到对应图片
-        switchCharacterImage(imageNumber);
-    } else {
-        console.log('mood不是有效数字，使用默认图片:', moodNumber);
-        // 如果不是数字或无效内容，切换到默认图片（1.png）
-        switchCharacterImage(1);
-    }
-}
-
-// 处理提取的【】内容（保留用于兼容性，但在新格式下不再使用）
-function handleExtractedBracketContents(extractedContents) {
-    // 确保 extractedContents 是数组
-    const contents = Array.isArray(extractedContents) ? extractedContents : [extractedContents];
-
-    // 处理每个提取的内容
-    for (const content of contents) {
-        console.log('处理【】内容:', content);
-
-        // 尝试解析为数字
-        const imageNumber = parseInt(content.trim());
-
-        if (!isNaN(imageNumber) && imageNumber > 0) {
-            // 如果是有效数字，切换到对应图片
-            switchCharacterImage(imageNumber);
-        } else {
-            console.log('【】内容不是有效数字，使用默认图片:', content);
-            // 如果不是数字或无效内容，切换到默认图片（1.png）
-            switchCharacterImage(1);
-        }
-    }
-}
-
-// 切换角色图片
-function switchCharacterImage(imageNumber) {
-    if (!currentCharacter) {
-        console.log('没有当前角色，无法切换图片');
-        return;
-    }
-
-    // 查找对应编号的图片
-    const targetImage = currentCharacterImages.find(img => img.number === imageNumber);
-
-    if (targetImage) {
-        console.log(`切换到图片 ${imageNumber}: ${targetImage.url}`);
-        const characterImage = document.getElementById('characterImage');
-        if (characterImage) {
-            characterImage.src = targetImage.url;
-        }
-    } else {
-        console.log(`未找到编号为 ${imageNumber} 的图片，使用默认图片`);
-        // 使用默认图片（1.png）
-        const defaultImage = currentCharacterImages.find(img => img.number === 1);
-        if (defaultImage) {
-            const characterImage = document.getElementById('characterImage');
-            if (characterImage) {
-                characterImage.src = defaultImage.url;
-            }
-        } else {
-            console.log('连默认图片都没有找到');
-        }
-    }
-}
-
-// 加载角色图片列表
-async function loadCharacterImages(characterId) {
-    if (!characterId) {
-        console.log('没有角色ID，无法加载图片');
-        return;
-    }
-
-    try {
-        console.log(`加载角色 ${characterId} 的图片列表`);
-
-        const response = await fetch(`/api/characters/${characterId}/images`);
-        const data = await response.json();
-
-        if (data.success) {
-            currentCharacterImages = data.images;
-            console.log(`成功加载 ${currentCharacterImages.length} 张图片:`, currentCharacterImages);
-
-            // 设置默认图片
-            const characterImage = document.getElementById('characterImage');
-            if (characterImage && data.default_image) {
-                characterImage.src = data.default_image;
-                console.log(`设置默认图片: ${data.default_image}`);
-            }
-        } else {
-            console.error('加载角色图片失败:', data.error);
-            currentCharacterImages = [];
-        }
-    } catch (error) {
-        console.error('加载角色图片时发生错误:', error);
-        currentCharacterImages = [];
-    }
-}
-
-// assistant内容更新时预请求音频
-function prefetchAudio(text, roleName, callback) {
-if (!text) {
-    if (callback) callback();
-    return;
-}
-if (audioCache[text]) {
-    // 命中缓存时，直接同步调用回调，避免异步等待
-    if (callback) callback();
-    return;
-}
-fetch('/api/tts', {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify({
-        text: text,
-        role: roleName || (currentCharacter ? currentCharacter.name : 'AI助手')
-    })
-})
-.then(response => {
-    if (!response.ok) return Promise.reject();
-    return response.blob();
-})
-.then(blob => {
-    if (blob && blob.size > 0) {
-        audioCache[text] = blob;
-    }
-    if (callback) callback();
-})
-.catch(() => {
-    if (callback) callback();
-});
-}
+// 暴露必要的函数给全局使用
+window.getCurrentCharacter = getCurrentCharacter;
+window.showOptionButtons = showOptionButtons;
+window.playAudio = (autoPlay = false) => playAudio(getCurrentCharacter(), autoPlay);
+window.stopCurrentAudio = stopCurrentAudio;
+window.showError = showError;
