+++ static/test.html
@@ -1,0 +1,134 @@
+<!DOCTYPE html>
+<html lang="zh-CN">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>CABM 模块化测试</title>
+    <style>
+        body {
+            font-family: Arial, sans-serif;
+            max-width: 800px;
+            margin: 0 auto;
+            padding: 20px;
+        }
+        .test-section {
+            margin: 20px 0;
+            padding: 20px;
+            border: 1px solid #ddd;
+            border-radius: 5px;
+        }
+        .success { color: green; }
+        .error { color: red; }
+        button {
+            margin: 5px;
+            padding: 10px 15px;
+            border: none;
+            border-radius: 3px;
+            cursor: pointer;
+        }
+        .primary { background: #007bff; color: white; }
+        .secondary { background: #6c757d; color: white; }
+    </style>
+</head>
+<body>
+    <h1>CABM JavaScript 模块化测试</h1>
+    
+    <div class="test-section">
+        <h2>模块加载测试</h2>
+        <div id="moduleStatus">检查中...</div>
+    </div>
+    
+    <div class="test-section">
+        <h2>DOM元素测试</h2>
+        <button id="testStartButton" class="primary">测试开始按钮</button>
+        <button id="testExitButton" class="secondary">测试退出按钮</button>
+        <div id="domStatus">等待测试...</div>
+    </div>
+    
+    <div class="test-section">
+        <h2>事件绑定测试</h2>
+        <button id="testEventButton" class="primary">测试事件</button>
+        <div id="eventStatus">等待测试...</div>
+    </div>
+
+    <script type="module">
+        // 测试模块加载
+        let moduleLoadSuccess = 0;
+        let moduleLoadTotal = 0;
+        
+        try {
+            moduleLoadTotal++;
+            const { startButton } = await import('./js/dom-elements.js');
+            if (startButton) moduleLoadSuccess++;
+            console.log('✓ dom-elements.js 加载成功');
+        } catch (e) {
+            console.error('✗ dom-elements.js 加载失败:', e);
+        }
+        
+        try {
+            moduleLoadTotal++;
+            const { showError } = await import('./js/ui-service.js');
+            if (showError) moduleLoadSuccess++;
+            console.log('✓ ui-service.js 加载成功');
+        } catch (e) {
+            console.error('✗ ui-service.js 加载失败:', e);
+        }
+        
+        try {
+            moduleLoadTotal++;
+            const { loadCharacters } = await import('./js/character-service.js');
+            if (loadCharacters) moduleLoadSuccess++;
+            console.log('✓ character-service.js 加载成功');
+        } catch (e) {
+            console.error('✗ character-service.js 加载失败:', e);
+        }
+        
+        try {
+            moduleLoadTotal++;
+            const { playAudio } = await import('./js/audio-service.js');
+            if (playAudio) moduleLoadSuccess++;
+            console.log('✓ audio-service.js 加载成功');
+        } catch (e) {
+            console.error('✗ audio-service.js 加载失败:', e);
+        }
+        
+        try {
+            moduleLoadTotal++;
+            const { sendMessage } = await import('./js/chat-service.js');
+            if (sendMessage) moduleLoadSuccess++;
+            console.log('✓ chat-service.js 加载成功');
+        } catch (e) {
+            console.error('✗ chat-service.js 加载失败:', e);
+        }
+        
+        // 更新状态
+        const moduleStatus = document.getElementById('moduleStatus');
+        if (moduleLoadSuccess === moduleLoadTotal) {
+            moduleStatus.innerHTML = `<span class="success">✓ 所有模块加载成功 (${moduleLoadSuccess}/${moduleLoadTotal})</span>`;
+        } else {
+            moduleStatus.innerHTML = `<span class="error">✗ 部分模块加载失败 (${moduleLoadSuccess}/${moduleLoadTotal})</span>`;
+        }
+        
+        // 测试DOM元素
+        document.getElementById('testStartButton').onclick = () => {
+            const domStatus = document.getElementById('domStatus');
+            try {
+                const startBtn = document.getElementById('startButton');
+                if (startBtn) {
+                    domStatus.innerHTML = '<span class="success">✓ startButton 元素存在于主页面</span>';
+                } else {
+                    domStatus.innerHTML = '<span class="error">✗ startButton 元素不存在</span>';
+                }
+            } catch (e) {
+                domStatus.innerHTML = `<span class="error">✗ DOM测试失败: ${e.message}</span>`;
+            }
+        };
+        
+        // 测试事件绑定
+        document.getElementById('testEventButton').onclick = () => {
+            const eventStatus = document.getElementById('eventStatus');
+            eventStatus.innerHTML = '<span class="success">✓ 事件绑定正常工作</span>';
+        };
+    </script>
+</body>
+</html>
