+++ templates/index.html
@@ -153,6 +153,6 @@
     
     <!-- JavaScript -->
     <script src="/static/js/stream_processor.js"></script>
-    <script src="/static/js/main.js"></script>
+    <script type="module" src="/static/js/main.js"></script>
 </body>
 </html>
